+++ wattsup/libftdi/cmake/has_ftdi_tcioflush.c
@@ -1,0 +1,6 @@
+#include <stdlib.h>
+#include <ftdi.h>
+
+void main(void) {
+  ftdi_tcioflush(NULL);
+}